Add task to uninstall packages
Used a bit different structure. Used a generic module `package` instead
of the `dnf` module. Also, used a `with_items` instead of listing
package names under `name`. When packages are listed in `with_items`,
then it is visible what particular items are being executed by the task.
That might be usefull in monitoring and debuging.