Don't add/remove yum repos with Ansible
Turns out, the `yum_repository` module in Ansible doesn't really work
good and as expected. Thus, I removed the task that adds RPM Fusion
repo. And added another task that runs a dnf command directly to disable
 the RPM Fusion repo that was added before and some other repos that
 appeared during the upgrade from Fedora v30 to v31.
VLC app was installed as a flatpak and thus was removed from the list of
 packages installed from RPM Fusion.