Remove dnf.conf file if it doesn't match condition
Using the `stat` module, it is possible to get info about the file and
use it later to set the condition. Here, I delete `/etc/dnf/dnf.conf`
file only if it is a regular file, because I want this file to be
exactly the same as the one in this playbook and thus I make a symlink.

Close #5